% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny.R
\name{add_mirtarbase_caption_shiny}
\alias{add_mirtarbase_caption_shiny}
\title{Add miRTarBase caption to plot}
\usage{
add_mirtarbase_caption_shiny(plot)
}
\arguments{
\item{plot}{Ggplot2 Object. Plot to add caption to.}
}
\value{
Plot with miRTarBase caption added.
}
\description{
Add miRTarBase caption to plot.
}
