% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_abstracts.R
\name{check_file_type_txt_shiny}
\alias{check_file_type_txt_shiny}
\title{Check if file type for upload ends in .txt}
\usage{
check_file_type_txt_shiny(file_name)
}
\arguments{
\item{file_name}{File name of file to upload.}
}
\value{
Boolean. If file name ends in .txt, it returns \code{TRUE}.
}
\description{
Check if file type for upload ends in .txt.
}
