% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_abstracts_compare.R
\name{common_mir_names_shiny}
\alias{common_mir_names_shiny}
\title{Get common miRNA names in ordered fashion}
\usage{
common_mir_names_shiny(df, mir_names_topic_ordered)
}
\arguments{
\item{df}{Data frame with two topics and miRNA names. Should be a result of
\code{upload_table_compare()}.}

\item{mir_names_topic_ordered}{Character vector. Containing reference miRNA
names. Product of \code{get_mir_names_shiny()}.}
}
\value{
Character vector of shared miRNA names in two topics for analysis.
}
\description{
Get common miRNA names from a data frame with a reference vector consisting
of miRNA names. The reference vector should have miRNA names in an ordered
fashion and should be a product of \code{get_mir_names_shiny()}.
}
