% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mir_count_shiny.R
\name{compare_mir_count_shared_shiny}
\alias{compare_mir_count_shared_shiny}
\title{Compare count of shared miRNAs in two topics}
\usage{
compare_mir_count_shared_shiny(df, mir, normalize, type)
}
\arguments{
\item{df}{Data frame with two topic names.}

\item{mir}{String or character vector. Specifies which miRNA count to
compare.}

\item{normalize}{Boolean. Specifies if count shall be normalized to the
relative amount of abstracts to all abstracts per topic, or if the absolute
count shall be used.}

\item{type}{String. Either \code{"standard"} or \code{"log2"}. Specifies if count
shall be compared on a standard scale or on a log2-scale.}
}
\value{
Plot with comparison of shared miRNA count.
}
\description{
Compare count of shared miRNAs in two topics.
}
