% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mir_count_shiny.R
\name{compare_mir_count_unique_shiny}
\alias{compare_mir_count_unique_shiny}
\title{Compare count of top unique miRNAs per topic}
\usage{
compare_mir_count_unique_shiny(df, top, normalize)
}
\arguments{
\item{df}{Data frame with two topic names}

\item{top}{Integer. Specifies of how many top unique miRNAs the count
shall be plotted.}

\item{normalize}{Boolean. Specifies if count shall be normalized to the
relative amount of abstracts to all abstracts per topic, or if the absolute
count shall be used.}
}
\value{
Plot with count of top unique miRNAs.
}
\description{
Compare count of top unique miRNAs per topic.
}
