% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mir_shiny.R
\name{count_mir_plot_shiny}
\alias{count_mir_plot_shiny}
\title{Plot top miRNA names}
\usage{
count_mir_plot_shiny(df, top = 10)
}
\arguments{
\item{df}{Data frame with miRNA names.}

\item{top}{Integer. Number of top miRNA counts to plot.}
}
\value{
Plot with top miRNA counts.
}
\description{
Plot top miRNA names. In contrast to miRetrieve,
the title is prefixed with the topic of the data frame.
}
