% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mir_shiny.R
\name{count_mir_table_shiny}
\alias{count_mir_table_shiny}
\title{Count miRNA names in table}
\usage{
count_mir_table_shiny(df)
}
\arguments{
\item{df}{Data frame with miRNA names.}
}
\value{
Data frame with counts.
}
\description{
Count miRNA names in table. In contrast to miRetrieve,
the second column is renamed in miRetrieveShiny.
}
