% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_terms_shiny.R
\name{count_mir_term_indicated_shiny}
\alias{count_mir_term_indicated_shiny}
\title{Count miRNAs of data frame with indicated terms}
\usage{
count_mir_term_indicated_shiny(df, term, threshold = 1, case = FALSE)
}
\arguments{
\item{df}{Data frame.}

\item{term}{Character vector. Terms to screen abstracts for.}

\item{threshold}{Integer. Specifies how often the term(s) must be
contained to be recognized.}

\item{case}{Boolean. Specifies if terms should be case sensitive or not.}
}
\value{
Plot with top 5 miRNAs for term subset
}
\description{
Count top 5 miRNAs of data frame with indicated terms for Shiny.
}
