% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny.R
\name{create_target_df_shiny}
\alias{create_target_df_shiny}
\title{Create target data frame}
\usage{
create_target_df_shiny(df_abstracts)
}
\arguments{
\item{df_abstracts}{Data frame containing PMIDs and abstracts.}
}
\value{
Target data frame. Data frame with four columns, namely
PMID, miRNA (as taken from miRTarBase), Target, and Topic
}
\description{
Create target data frame for Shiny. Joins data frame
with abstracts with miRTarBase 8.0 on its PMIDs and
selects PMIDs, miRNAs as taken from miRTarBase, Targets, and
Topic.
}
\details{
These are not the same miRNAs as extracted while uploading the abstracts!
The miRNAs are taken from miRTarBase, where only the miRNA stem (e.g. "miR-200"
instead of "hsa-miR-200b-5p") is extracted.
}
