% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_shiny.R
\name{get_min_year_shiny}
\alias{get_min_year_shiny}
\title{Get min publication year}
\usage{
get_min_year_shiny(df, col.year = Year)
}
\arguments{
\item{df}{Data frame to extract publication year from.}

\item{col.year}{Symbol. Column containing publication year.}
}
\value{
integer
}
\description{
Get lowest publication year in a data frame
}
