% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mir_names.R
\name{get_mir_names_shiny}
\alias{get_mir_names_shiny}
\title{Get miRNA names from a data frame}
\usage{
get_mir_names_shiny(df, col.mir = miRNA)
}
\arguments{
\item{df}{Data frame with miRNA names.}

\item{col.mir}{Symbol. Column containing miRNA names.}
}
\value{
Character vector.
}
\description{
Get miRNA names as a vector from a data frame. miRNA names
are sorted in order of their count, e.g. the miRNA with the highest
frequency is at the beginning.
}
