% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny_compare.R
\name{get_shared_target_names_shiny}
\alias{get_shared_target_names_shiny}
\title{Get names of shared targets between two topics}
\usage{
get_shared_target_names_shiny(df_target_compare, target_name_ref)
}
\arguments{
\item{df_target_compare}{Data frame with targets for two topics,
based on miRTarBase 8.0}

\item{target_name_ref}{Character vector. Reference target character vector
with target names sorted by reference. Product of
\code{get_target_target_names_shiny()}.}
}
\value{
Character vector with shared target names for two topics, sorted by
frequency based on reference character vector.
}
\description{
Get names of shared targets between two topics.
}
