% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny_compare.R
\name{get_target_mir_names_compare_shiny}
\alias{get_target_mir_names_compare_shiny}
\title{Extract miRNA names for target operations to compare}
\usage{
get_target_mir_names_compare_shiny(
  df_abstracts,
  df_target,
  mir_names_topic_ordered
)
}
\arguments{
\item{df_abstracts}{Data frame containing PMIDs and abstracts of two topics.}

\item{df_target}{Data frame with miRNA targets based on miRTarBase 8.0,
usually created by \code{create_target_df_shiny()}.}

\item{mir_names_topic_ordered}{Character vector. Containing reference miRNA
names. Product of \code{get_mir_names_shiny()}.}
}
\value{
Character vector with miRNA names from \code{df_abstracts} with targets in
miRTarBase 8.0, sorted by count frequeny in \code{df_abstracts}.
}
\description{
Extract miRNA names for target operations to compare, sorted by frequency
}
