% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny.R
\name{get_target_target_names_shiny}
\alias{get_target_target_names_shiny}
\title{Get names of miRNA targets based on target df}
\usage{
get_target_target_names_shiny(df_target)
}
\arguments{
\item{df_target}{Target data frame.}
}
\value{
Character vector with targets names sorted by count frequency.
}
\description{
Get names of miRNA targets based on target df, sorted by count
frequency.
}
