% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mir_names.R
\name{get_topic_name_shiny}
\alias{get_topic_name_shiny}
\title{Get topic name of data frame.}
\usage{
get_topic_name_shiny(df, col.topic = Topic)
}
\arguments{
\item{df}{Data frame with topic names.}

\item{col.topic}{Symbol. Column containing topic name.}
}
\value{
String with the topic name
}
\description{
Get topic name of Topic in data frame.
}
