% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_subset_mir_shiny.R
\name{indicate_mir_shiny}
\alias{indicate_mir_shiny}
\title{Indicate miRNA in data frame}
\usage{
indicate_mir_shiny(df, indicate.mir = NULL, col.mir = miRNA)
}
\arguments{
\item{df}{Data frame.}

\item{indicate.mir}{String or character vector. miRNAs to indicate.}

\item{col.mir}{Symbol. Column containing miRNA names.}
}
\value{
Data frame.
}
\description{
Indicates presence of miRNA in abstract with "Yes"/"No" column. Different
column name as opposed to miRetrieve.
}
