% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_biomarker_shiny.R
\name{plot_mir_biomarker_shiny}
\alias{plot_mir_biomarker_shiny}
\title{Plot top likely miRNA biomarker}
\usage{
plot_mir_biomarker_shiny(df, threshold = 1)
}
\arguments{
\item{df}{Data frame.}

\item{threshold}{Integer. Threshold to distinguish biomarker abstracts.}
}
\value{
Plot with top most likely biomarkers.
}
\description{
Plot top likely miRNA biomarker. Biomarker abstracts are identified
with calculate_score and a threshold, rest of abstracts are discarded.
miRNAs are counted.
}
