% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_development_shiny.R
\name{plot_mir_development_shiny}
\alias{plot_mir_development_shiny}
\title{Plot development of miRNAs in abstracts}
\usage{
plot_mir_development_shiny(df, mir, start_year, end_year)
}
\arguments{
\item{df}{Data frame with miRNAs and publication years.}

\item{mir}{String or character vector. Specifies which miRNAs to plot.}

\item{start_year}{Integer. Start year.}

\item{end_year}{Integer. End year.}
}
\value{
Line plot.
}
\description{
Plot development of miRNAs in abstracts.
}
