% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mir_terms_shiny.R
\name{plot_mir_terms_shiny}
\alias{plot_mir_terms_shiny}
\title{Plot miRNA-terms association}
\usage{
plot_mir_terms_shiny(df, mir, top = 20, token = "words",
  normalize = TRUE)
}
\arguments{
\item{df}{Data frame.}

\item{mir}{String. Specifies which miR to plot terms for.}

\item{top}{Integer. Number of top terms to plot.}

\item{token}{String. Either 'words' or '2-grams'.}

\item{normalize}{Boolean. Specifies if count shall be normalized.}
}
\value{
Plot.
}
\description{
Plot miRNA-terms association. Less arguments than in miRetrieve,
short cut to tokenization.
}
