% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny.R
\name{plot_target_count_shiny}
\alias{plot_target_count_shiny}
\title{Plot top validated targets}
\usage{
plot_target_count_shiny(df_target, top)
}
\arguments{
\item{df_target}{Target data frame.}

\item{top}{Integer. Number of top targets to plot.}
}
\value{
Bar plot with top targets.
}
\description{
Plot top validated targets as a bar plot.
}
