% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny.R
\name{plot_target_mir_scatter_mir_shiny}
\alias{plot_target_mir_scatter_mir_shiny}
\title{Plot target-miR scatter plot based on miRNAs}
\usage{
plot_target_mir_scatter_mir_shiny(df_target, mir)
}
\arguments{
\item{df_target}{Target data frame, created by \code{create_target_df_shiny()}.}

\item{mir}{Character vector. miRNAs to plot.}
}
\value{
Scatter plot with target-miR interactions based on miRNAs.
}
\description{
Plot target-miR scatter plot based on miRNAs
}
