% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_shiny.R
\name{plot_target_mir_scatter_target_shiny}
\alias{plot_target_mir_scatter_target_shiny}
\title{Plot target-miR scatter plot based on targets}
\usage{
plot_target_mir_scatter_target_shiny(df_target, target)
}
\arguments{
\item{df_target}{Target data frame, created by \code{create_target_df_shiny()}.}

\item{mir}{Character vector. Targets to plot.}
}
\value{
Scatter plot with target-miR interactions based on targets.
}
\description{
Plot target-miR scatter plot based on targets
}
