% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preloaded_pubmed_doc.R
\docType{data}
\name{preloaded_pubmed}
\alias{preloaded_pubmed}
\title{PubMed Abstracts of Malignant Melanoma}
\format{
: A data frame
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/}
}
\usage{
preloaded_pubmed
}
\description{
A dataset with miRNA names extracted from abstracts of Malignant Melanoma as
of October 1st, 2020.
Threshold = 2, extract_names = FALSE
}
\keyword{datasets}
