% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_abstracts.R
\name{set_topic_shiny}
\alias{set_topic_shiny}
\title{Set topic for data frame}
\usage{
set_topic_shiny(topic_input, topic_set)
}
\arguments{
\item{topic_input}{Source of topic string.}

\item{topic_set}{String. What the topic shall be named if topic_input == "".}
}
\value{
String with topic name.
}
\description{
Set topic for data frame. No idea if this works, but Shiny does not want to
accept my ideas.
If topic_input == "", then topic = topic_set. If topic_input != "", then
topic = topic_input.
}
