% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_mir_perc_shiny.R
\name{sort_mir_perc_shiny}
\alias{sort_mir_perc_shiny}
\title{Helper - Sort miRNAs of two topic by average relative count}
\usage{
sort_mir_perc_shiny(df, col.mir = miRNA, col.pmid = PMID,
  col.topic = Topic)
}
\arguments{
\item{df}{Data frame after extracting miRNAs}

\item{col.mir}{Symbol. Column containing miRNA names}

\item{col.pmid}{Symbol. Column containing PubMed-IDs.}

\item{col.topic}{Symbol. Column containing topic names.}
}
\value{
miRNA vector sorted with relative count.
}
\description{
Sort relative miRNA count of two topics by average of their
relative count. Mainly created to sort the miRNAs that are often
reported in both topics to the top for the Shiny application.
}
