% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_columns_shiny.R
\name{subset_columns_research_shiny}
\alias{subset_columns_research_shiny}
\title{Subset data frame for abstracts of research articles}
\usage{
subset_columns_research_shiny(df, col.lang = Language, col.type = Type,
  subset = TRUE)
}
\arguments{
\item{df}{Data frame containing article types.}

\item{col.lang}{Symbol. Column containing Language type.}

\item{col.type}{Symbol. Column containing articles types.}

\item{subset}{Boolean. If \code{subset = TRUE}, subset for abstracts of
original research articles.}
}
\value{
Data frame containing abstracts of research articles only.
}
\description{
Subset data frame for abstracts of research articles only. Special function
for miRetrieve Shiny.
}
\details{
Optionally, subset data frame for abstracts of research articles only.
At the same time, abstracts from other article types such as \emph{Review},
\emph{Letter}, etc. are dropped.
Most importantly, columns calles "Type" and "Language" are dropped
from the data frame after reading it into R.
Special Shiny function.
}
\seealso{
\code{\link[=subset_review]{subset_review()}}, \code{\link[=subset_year]{subset_year()}}
}
\concept{subset functions}
