% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_subset_mir_shiny.R
\name{subset_mir_shiny}
\alias{subset_mir_shiny}
\title{Subet data frame for miRNA}
\usage{
subset_mir_shiny(df, mir = NULL, col.mir = miRNA)
}
\arguments{
\item{df}{Data frame.}

\item{mir}{String or character vector. miRNAs to subset for.}

\item{col.mir}{Symbol. Column containing miRNA names.}
}
\value{
Data frame.
}
\description{
Subet data frame for miRNA
}
