% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_abstracts.R
\name{upload_abstracts_error_shiny}
\alias{upload_abstracts_error_shiny}
\title{Upload abstracts and catch errors}
\usage{
upload_abstracts_error_shiny(...)
}
\arguments{
\item{...}{Arguments passed to \code{upload_abstracts_shiny()}.}
}
\value{
Either tibble if the functions works, or string with error message
if a wrong file was used.
}
\description{
Catch error messages when uploading abstracts and return a character value.
}
