% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_abstracts.R
\name{upload_abstracts_shiny}
\alias{upload_abstracts_shiny}
\title{Upload abstracts for Shiny}
\usage{
upload_abstracts_shiny(medline_file, threshold = 1, topic = NULL,
  subset_research = TRUE, extract_letters = FALSE)
}
\arguments{
\item{medline_file}{Medline-file.}

\item{topic}{String. Specifies topic name.}

\item{subset_research}{Boolen. If TRUE, only abstracts of original research
articles are kept.}

\item{extract_letters}{Boolean. If TRUE, letters are extracted from miRNA
names (e.g. miR-23a).}

\item{threshold.}{Integer. Specifies how often a miRNA must be mentioned
in Abstracts to be extracted.}
}
\value{
Uploaded data frame.
}
\description{
Upload abstracts for Shiny. Unlike in miRetrieve,
columns "Type" and "Language" are dropped in miRetrieve Shiny after
upload.
}
