% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_abstracts_compare.R
\name{upload_table_compare_shiny}
\alias{upload_table_compare_shiny}
\title{Upload second data set to compare}
\usage{
upload_table_compare_shiny(file_compare, df_present,
  topic_compare = "Topic_compare", threshold = 1,
  subset_research = TRUE, extract_letters = FALSE)
}
\arguments{
\item{file_compare}{String. File path to upload.}

\item{df_present}{Data frame with extracted miRNAs and topic name. This is the
reference data frame that the newly uploaded file is joined to.}

\item{topic_compare}{String. Name of topic for the new data frame.}

\item{threshold}{Integer. Specifies how often a miRNA must be in an
abstract to be extracted.}

\item{subset_research}{Boolean. Specifies if only abstracts of original research
articles shall be kept.}

\item{extract_letters}{Boolean. Specifies if miRNA shall be extracted with
trailing letters (e.g. miR-23a).}
}
\value{
Data frame with miRNAs extracted and two topics
}
\description{
Upload a second data set to compare. This data set is automatically joined to
\code{df_present}.
}
